backend-django/requirements.txt
django==5.2.*
djangorestframework==3.14.*
django-cors-headers==4.4.*
python-decouple==3.8
psycopg2-binary==2.9.*
grpcio>=1.62.0 
protobuf>=5.26.0
grpcio-tools>=1.62.0
django-grpc==0.1.1 
djangorestframework-simplejwt==5.3.*